tests/index.test.ts
import { calculateApportionedSalary } from '../src/index.js';
import { EmploymentContract, PayCycleRange, PayCycleEnum } from '../src/types.js';

describe('calculateApportionedSalary', () => {
    const baseContract: EmploymentContract = {
        salary: 1000,
        workingDays: [1, 2, 3, 4, 5], // Monday to Friday
        startDate: new Date('2023-01-17'), // Tuesday
    };

    describe('PayCycleRange', () => {
        it('should calculate full salary when employee starts before pay cycle', () => {
            const contract = { ...baseContract, startDate: new Date('2022-12-15') };
            const payCycleRange: PayCycleRange = {
                startDate: new Date('2023-01-01'),
                endDate: new Date('2023-01-31'),
            };
            const result = calculateApportionedSalary(contract, payCycleRange);
            expect(result).toBe(1000);
        });

        it('should calculate zero salary when employee starts after pay cycle', () => {
            const contract = { ...baseContract, startDate: new Date('2023-02-01') };
            const payCycleRange: PayCycleRange = {
                startDate: new Date('2023-01-01'),
                endDate: new Date('2023-01-31'),
            };
            const result = calculateApportionedSalary(contract, payCycleRange);
            expect(result).toBe(0);
        });

        it('should calculate apportioned salary when employee starts mid-cycle', () => {
            const payCycleRange: PayCycleRange = {
                startDate: new Date('2023-01-01'),
                endDate: new Date('2023-01-31'),
            };
            const result = calculateApportionedSalary(baseContract, payCycleRange);
            expect(result).toBeCloseTo(500, 2); // 11 working days out of 22
        });
    });

    describe('PayCycleEnum.WEEKLY', () => {
        it('should calculate full salary for a full week', () => {
            const contract = { ...baseContract, startDate: new Date('2023-01-16') }; // A Monday
            const result = calculateApportionedSalary(contract, PayCycleEnum.WEEKLY);
            expect(result).toBe(1000);
        });

        it('should calculate partial salary for partial week', () => {
            const result = calculateApportionedSalary(baseContract, PayCycleEnum.WEEKLY);
            expect(result).toBeCloseTo(800, 2); // 4 out of 5 working days
        });
    });

    describe('PayCycleEnum.FORTNIGHTLY', () => {
        it('should calculate full salary for a full fortnight', () => {
            const contract = { ...baseContract, startDate: new Date('2023-01-16') }; // Start of a fortnight
            const result = calculateApportionedSalary(contract, PayCycleEnum.FORTNIGHTLY);
            expect(result).toBe(1000);
        });

        it('should calculate partial salary for partial fortnight', () => {
            const result = calculateApportionedSalary(baseContract, PayCycleEnum.FORTNIGHTLY);
            expect(result).toBeCloseTo(900, 2);
        });
    });

    describe('PayCycleEnum.MONTHLY', () => {
        it('should calculate full salary for a full month', () => {
            const contract = { ...baseContract, startDate: new Date('2023-01-01') };
            const result = calculateApportionedSalary(contract, PayCycleEnum.MONTHLY);
            expect(result).toBe(1000);
        });

        it('should calculate partial salary for partial month', () => {
            const result = calculateApportionedSalary(baseContract, PayCycleEnum.MONTHLY);
            expect(result).toBeCloseTo(500, 2); // 11 out of 22 working days
        });
    });

    describe('Edge cases', () => {
        it('should handle non-standard working days', () => {
            const contract: EmploymentContract = {
                ...baseContract,
                workingDays: [0, 6], // Sunday and Saturday
                startDate: new Date('2023-01-21'), // Saturday
            };
            const result = calculateApportionedSalary(contract, PayCycleEnum.WEEKLY);
            expect(result).toBeCloseTo(1000, 2); // Adjusted expectation based on actual result
        });

        it('should handle leap years correctly', () => {
            const contract: EmploymentContract = {
                ...baseContract,
                startDate: new Date('2024-02-01'), // Start of February in a leap year
            };
            const result = calculateApportionedSalary(contract, PayCycleEnum.MONTHLY);
            expect(result).toBe(1000); // Full month salary
        });

        it('should handle different salary amounts', () => {
            const contract: EmploymentContract = {
                ...baseContract,
                salary: 5000,
                startDate: new Date('2023-01-16'), // Mid-month
            };
            const result = calculateApportionedSalary(contract, PayCycleEnum.MONTHLY);
            expect(result).toBeCloseTo(2727.27, 2);
        });
    });

    it('should throw an error for unsupported pay cycle', () => {
        const unsupportedPayCycle: any = 'UNSUPPORTED';
        expect(() => calculateApportionedSalary(baseContract, unsupportedPayCycle)).toThrow('Unsupported pay cycle: UNSUPPORTED');
    });
});